projects/skills-hub-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skills-hub"
version = "1.0.0"
description = "Hermes Skills Hub CLI — browse and search Hermes skills"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Hermes Agent"}
]
dependencies = [
    "pyyaml",
]

[project.scripts]
skills-hub = "skills_hub.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools]
packages = ["skills_hub"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
